Change the formatters/encoders/layouts to be composed of multiple JsonProviders instead of doing all the work themselves.
Each JsonProvider is responsible for writing a specific part of the event
to the JsonGenerator used to create the JSON output.

This allows greater flexibility for adding new data to the JSON output.

LogstashLayout/LogstashEncoder (and their access event equivalents)
remain fully backwards compatible.

New CompositeJsonEncoder/CompositeJsonLayout classes
(and subclasses for LoggingEvents and AccessEvents) have been created,
and are used as superclasses for the LogstashLayout/Encoder/Formatter.

These new composite classes are composed of JsonProviders,
and can be configured/used independently of the LogstashLayout/Encoder/Formatter classes.

The LogstashLayout/Encoder/Formatter classes can now be thought of as
"pre-defined" composites, and will continue to be supported.
However, not all new features may be exposed through them.
Complex configuration may require the use of the composite classes instead.