Jackson 3.0 initial support
Based on Jackson 3.0.0-SNAPSHOT (as of today).

Jackson 3.0 made some backwards incompatible changes.
Therefore, there are some corresponding backwards incompatible changes in logstash-logback-encoder.
See changes_to_support_jackson_3.0.md for all of the changes made.

Due to the nature of the Jackson changes, it will be impossible to support
both Jackson 2.x and 3.x in the same version of logstash-logback-encoder.