Attempt to fix unstable AsyncDisruptorAppenderTest#testEventHandlerCalled test case (#623)
* Rewrite test case with concrete test classes instead of Mockito

I could not reproduce the issue locally. I suspect some nasty concurrent issues with Mockito... Decided to rewite the test case entirely with concrete classes instead of Mockito. Hope it helps...

* Async processing is done when exceptionHandler has processed the exception

We need to wait for the exception handler to have processed the exception to determine when async processing is completed. Waiting for the event handler is not enough...