Javadoc, remove `#finalize()`, do not recycle after exception
Do not rely on `#finalize()` anymore to close the underlying JsonGenerator. The JsonGenerator is *not* closed anymore when the ReusableJsonFormatter is disposed. This means that the internal buffers allocated by the generator are not returned to Jackson's internal pool which frees even more memory.

Do not recycle a ReusableJsonFormatter after the underlying JsonFormatter threw an exception. The underlying JsonGenerator may be in an unsafe state.