LogstashMarker refactoring (#636)
* Lazy creation of Marker referenceList

Most Logstash marker don't have any children/reference at all. Creating the referenceList upfront is a waste of memory and additional pressure on the GC. This commit modifies the implementation to create the reference list when the first reference is added.

* Remove copyright from original SLF4J implementation

This class is now our own implementation - remove copy of SLF4J copyright notice

* Reduce visibility of LogstashBasicMarker

fixes #637

* Code cleanup and additional javadoc

* Be consistent in usage of `this` to reference instance fields

* Review equals/hashCode on EmptyLogstashMarker and add obj.getClass() in comparison

* Accept `null` value in contains()

contains() should accept null values and return false in this case. Calling code will not have to test for nullability before using the method and still have a consistent result

* Add test cases for LogsatshBasicMarker