Delegate recycling to LogEvent itself
Sub-classes of the AsyncDisruptorAppender may provide their own LogEventFactory and LogEventTranslator. They may also work with an extended version of the base LogEvent with additional fields. Recycling these instances may therefore require more than simply setting the `event` field back to null before the instance is returned to the RingBuffer.
This commit adds a `LogEvent#recycle()` method whose job is to perform the needed tasks on the log event before it can be reused. It is called by the `EventClearingEventHandler` wrapper after the event is processed by the EventHandler.