Refactor AsyncAppenderDisruptor event handling (#651)
* Use createEventHandler() to provide the EventHandler instead of 'setEventHandler()'

Introduce `#createEventHandler()` method to provide the EventHandler instead of a simple setter method.
Reasons include:
- providing an EventHandler is mandatory. Using an abstract method makes obvious one is required whereas a simple setter supposes it may be optional (until start() is called... which is at runtime, and therefore too late).
- inheriting classes must call `#setEventHandler()` in their constructor - the base class does not have control as to when the handler is actually created. Also, if a class wants to extend `AsyncDisruptorAppender` to provide its own implementation, a first handler will be created and set by AsyncDisruptorAppender's constructor before it is overridden by the one provided by the extending class. Using a factory method solves that issue.
- sub-classes may decide if a new instance of their handler must be used every time the appender is start/stop or if the same instance can be reused.

* Make LogEventFactory class "protected" instead of "private"

The protected method `#setEventFactory(LogEventFactory)` cannot be used by extending classes if the `LogEventFactory` static inner class is *private*...

* Delegate recycling to LogEvent itself

Sub-classes of the AsyncDisruptorAppender may provide their own LogEventFactory and LogEventTranslator. They may also work with an extended version of the base LogEvent with additional fields. Recycling these instances may therefore require more than simply setting the `event` field back to null before the instance is returned to the RingBuffer.
This commit adds a `LogEvent#recycle()` method whose job is to perform the needed tasks on the log event before it can be reused. It is called by the `EventClearingEventHandler` wrapper after the event is processed by the EventHandler.