Don't repeat exception cause message when attaching the exception to the status
The default "OnConsoleStatusListener" already appends the cause's message when printing tth status on the console. Addining it in the original message will cause duplicates on the console.
Those not relying on the "OnConsoleStatusLisener" should simply retrieve additional information from the status  throwable if any (what they probaby already do for ErrorStatus anyway...)