Reschedule maven-check-style plugin during phase "process-sources" together with animal sniffer
Project validation steps should be executed as early as possible during the build. We must somehow guarantee that the project conforms to the expectations before executing any tests or producing final artifacts. Checkstyle is one of the validation criteria and as much should be executed as soon as possible.

Also make the execution phase explicit in the POM even if the phase is the same as the default. This helps reading the POM and have a quick view of when each plugin is executed without having to remember their defaults.