Delegate to `JsonGenerator@writeObject()` when value is not a simple or type or JsonNode (#704)
Calling JsonNode#serialize() with a null SerializerProvider is causing issue for Array and Object types. These types rely on the SerializerProvider to serialize null values for instance...

The JsonWritingUtils#writeObject() method is enhanced to handle as much types as possible before delegating to JsonGenerator#writeObject() if it cannot handle the value.