FastISOTimestampFormatter as a fast alternative to Java's DateTimeFormatter for standard ISO formats (#711)
The FastISOTimestampFormatter is a fast alternative to Java's DateTimeFormatter for the most commonly used ISO formats, i.e.:
- ISO_OFFSET_DATE_TIME
- ISO_ZONED_DATE_TIME
- ISO_LOCAL_DATE_TIME
- ISO_DATE_TIME
- ISO_INSTANT

Instead of computing all date/time fields every time it is invoked like the standard DateTimeFormatter does, the fast formatter remembers the previous formatted value and re-uses it by replacing the second and millis parts when the next timestamp to format is within the same minute. If not, it delegates to the enclosed DateTimeFormatter to get a fresh new formatted value.