Custom JsonGenerator with optimized writeObject() for simple object types
Wrap the JsonGenerator with an optimized version of the writeObject() method that tries to call the appropriate write() method for the giveb untype Object and delegate to the underlying JsonGenerator as fallback.

Without this optimisation the writeObject() delegates to the underlying Codec which is less effective for simple types (String, Numbers, Boolean, byte[], JsonNode, etc) both in terms of CPU and temporary memory allocations.

Until now only the AbstractJsonPatternParser was using this optimisation. This commit moves it directly inside the JsonGenerator so it can benefit to any JsonProvider.