ThreadLocalHolder: Optimize threadValues
Use a concurrent hash set instead of a CopyOnWriteArrayList.
threadValues.remove(...) is O(n) and also locks the
threadValues object which means that any other threads
that operate on it will block.

If there are many threads being created and killed,
the overhead of using CopyOnWriteArrayList
becomes significant.

Using as concurrent hash set (via ConcurrentHashMap),
add and remove become O(1) and avoids locking.