Add ability to customize the PrettyPrinter
Add ability to customize the PrettyPrinter used by PrettyPrintingJsonGeneratorDecorator.

Specifically, add rootSeparator and spacesInObjectEntries fields.

Change default root value separator used by the PrettyPrinter to an empty string to match the default rootValueSeparator used by the formatters.
Previously, the default was a space (" ").  So this will slightly change the output when the PrettyPrintingJsonGeneratorDecorator is used.
To revert to the old behavior, configure the decorator like this:

<jsonGeneratorDecorator class="net.logstash.logback.decorate.PrettyPrintingJsonGeneratorDecorator">
	<rootSeparator>[SPACE]</rootSeparator>
</jsonGeneratorDecorator>

Requested in discussion #906