Fix not-JSON generators
non-JSON generators were broken in 7.0 when a call to generator.setRootValueSeparator was introduced.  The non-JSON generators throw UnsupportedOperationException when setRootValueSeparator is invoked.

Handle the UnsupportedOperationException properly.

Also moved disabling the FLUSH_PASSED_TO_STREAM generator feature to prior to decoration, because the YAML generation needs this feature enabled to work properly.

Introduced basic tests for non-JSON generators to catch the above problems in the future.

Fixes #919